requirements.txt
python-binance==1.0.15
pandas==1.3.5

parameterized==0.8.1
PyYAML==6.0